# server/openapi.yaml
...
openapi: 3.1.0
info:
  title: GitStars Service API
  version: 1.0.0
  description: |
    Contract for the GitStars service (local/self-hosted backend, Hono + SQLite).

    Conventions (ADR-0006):
    - Every response carries `X-Request-Id` (client-supplied or server-generated).
    - Mutations marked `Idem` honor `Idempotency-Key`; replays return the stored
      response without re-applying (ADR-0004 D2).
    - Conditional writes use `If-Match: "<entityId>:<version>"`; stale base → 409
      VERSION_CONFLICT with `details.current`.
    - All errors use the standard envelope `{ code, message, details?, request_id }`
      with the machine-decidable codes cataloged below.
    - Compatibility: additive changes (new optional fields/entities) do not bump
      `PROTOCOL_VERSION`; breaking changes bump it and observe an N-1 deprecation
      window, after which older clients receive 426 STALE_CLIENT (ADR-0004 D5).
servers:
  - url: /api
...
security:
  - sessionCookie: []
paths:
...
  /changes:
    get:
      tags: [sync]
      summary: Incremental change feed (ADR-0004 D2/D6)
      description: >
        Ordered per-user feed with monotonic `seq` cursor. Entries carry an
        additive optional `data` payload with the row's current committed state so
        pull clients can materialize their replica without per-entity refetches;
        hard-deleted rows have `data: null` (client deletes by entityId) and
        repository_tag deletions reconstruct the composite key from entityId.
      parameters:
        - {
            name: since,
            in: query,
            schema: { type: integer, minimum: 0 },
            description: last seen seq,
          }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 500 },
          }
      responses:
        "200":
          description: Change page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChangeFeed" }
        "410":
          description: >
            CURSOR_INVALID — the cursor is ahead of the feed tail (e.g. server-side
            database reset). The client must re-pull from 0; queued local mutations
            are unaffected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /library:
    get:
      tags: [library]
      summary: List saved repositories with filter/sort
      description: >
        Filter/sort per ADR-0006 D3: `provider`, `tag`, `status`,
        `sort=added_at|stars|name`, `order=asc|desc`.
      responses:
        "200":
          description: Saved repositories (joined with normalized repository metadata)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SavedRepository" }
    post:
      tags: [library]
      summary: Save a repository by reference (identity resolution server-side)
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [providerType, canonicalKey]
              properties:
                providerType: { type: string }
                canonicalKey: { type: string }
                host: { type: string }
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavedRepository" }
        "400": { $ref: "#/components/responses/Error400" }
...
  /tags:
    get:
      tags: [tags]
      summary: List own tags
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tag" }
    post:
      tags: [tags]
      summary: Create tag (idempotent by user+name)
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "201":
          description: Created (200 when the name already exists)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "400": { $ref: "#/components/responses/Error400" }
...
  /lists:
    get:
      tags: [lists]
      summary: List own lists
      responses:
        "200":
          description: Lists
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/List" }
    post:
      tags: [lists]
      summary: Create list
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/List" }
        "400": { $ref: "#/components/responses/Error400" }
        "409": { $ref: "#/components/responses/Error409" }
...
  /lists/{id}/publication:
    post:
      tags: [sharing]
      summary: >
        Publish or republish a list (share link). Republishing an active
        publication refreshes the sanitized snapshot in place and bumps
        snapshot_version; a revoked publication is reactivated. Rate limited to
        10 requests per user per hour.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      responses:
        "201":
          description: Publication created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publication" }
        "200":
          description: Publication refreshed (republish/reactivate)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publication" }
        "404": { $ref: "#/components/responses/Error404" }
        "409": { $ref: "#/components/responses/Error409" }
        "429": { $ref: "#/components/responses/Error429" }
    get:
      tags: [sharing]
      summary: Owner-only publication status (404 when never published)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Publication status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publication" }
        "404": { $ref: "#/components/responses/Error404" }
    delete:
      tags: [sharing]
      summary: Revoke the share link (unpublish; the list itself is untouched)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                {
                  type: object,
                  required: [ok],
                  properties: { ok: { type: boolean } },
                }
        "404": { $ref: "#/components/responses/Error404" }
  /lists/{id}/hub:
    put:
      tags: [sharing]
      summary: Opt the active publication into (or out of) the public Hub catalog
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [optIn]
              properties:
                optIn: { type: boolean }
      responses:
        "200":
          description: Hub visibility updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, hubOptIn]
                properties:
                  ok: { type: boolean }
                  hubOptIn: { type: boolean }
        "400": { $ref: "#/components/responses/Error400" }
        "404": { $ref: "#/components/responses/Error404" }
        "409": { $ref: "#/components/responses/Error409" }
...
  /preferences:
    get:
      tags: [preferences]
      summary: Read preference singleton (ETag)
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preferences" }
    put:
      tags: [preferences]
      summary: Shallow field-level merge, then version bump (ADR-0004 D4)
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                data: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Merged preferences
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preferences" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: gitstars_session
    adminBearer:
      type: http
      scheme: bearer
      description: ADMIN_TOKEN bearer credential for admin endpoints (M5)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, format: uuid }
      description: Replay-safe mutation key (ADR-0004 D2)
...
  responses:
    Error400:
      description: VALIDATION — bad input
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
...
    Error404:
      description: NOT_FOUND
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
    Error409:
      description: VERSION_CONFLICT (details.current) or CONFLICT (uniqueness)
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
    Error429:
      description: RATE_LIMITED (details.resetAt)
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - VERSION_CONFLICT
            - CONFLICT
            - STALE_CLIENT
            - RATE_LIMITED
            - CURSOR_INVALID
            - PROVIDER_NOT_IMPLEMENTED
            - PROVIDER_ERROR
            - OAUTH_NOT_CONFIGURED
            - OAUTH_EXCHANGE_FAILED
            - OAUTH_STATE_INVALID
            - USER_FETCH_FAILED
            - DEV_LOGIN_DISABLED
            - ADMIN_NOT_CONFIGURED
            - INTERNAL_ERROR
        message: { type: string }
        details: { type: object, additionalProperties: true }
        request_id: { type: string }
...
    SavedRepository:
      type: object
      required: [id, version, etag, status, repository]
      properties:
        id: { type: string }
        version: { type: integer }
        etag: { type: string }
        status: { type: string }
        note: { type: string }
        aiSummary: { type: string }
        aiTags: { type: array, items: { type: string } }
        addedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        repository:
          type: object
          required: [id, providerType, name, webUrl]
          properties:
            id: { type: string }
            providerType: { type: string }
            host: { type: string }
            remoteId: { type: string }
            canonicalKey: { type: string }
            name: { type: string }
            namespacePath: { type: string }
            webUrl: { type: string }
            description: { type: string }
            visibility: { type: string }
            primaryLanguage: { type: string }
            starsCount: { type: integer }
            forksCount: { type: integer }
            status: { type: string }
        tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
...
    Tag:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
    List:
      type: object
      required: [id, name, version, etag]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        version: { type: integer }
        etag: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        itemCount: { type: integer }
...
    Preferences:
      type: object
      required: [data, version]
      properties:
        data: { type: object, additionalProperties: true }
        version: { type: integer }
        etag: { type: string }
    ChangeEntry:
      type: object
      required: [seq, entityType, entityId, op, version, createdAt]
      properties:
        seq: { type: integer }
        entityType:
          type: string
          enum:
            - saved_repository
            - list
            - list_item
            - tag
            - repository_tag
            - preference
        entityId: { type: string }
        op: { type: string, enum: [created, updated, deleted] }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        data:
          description: >
            Sanitized current-row payload (additive, ADR-0004 D5); null when the
            row is hard-deleted. Tombstoned rows include deletedAt.
    ChangeFeed:
      type: object
      required: [changes, nextCursor, hasMore, protocolVersion]
      properties:
        changes:
          type: array
          items: { $ref: "#/components/schemas/ChangeEntry" }
        nextCursor: { type: integer }
        hasMore: { type: boolean }
        protocolVersion: { type: integer }
...
    Publication:
      type: object
      description: >
        Owner view of a list publication. shareId is a 26-char Crockford base32
        public handle (never a UUID, never the internal list id).
      required:
        [
          shareId,
          status,
          snapshotVersion,
          title,
          repositoryCount,
          updatedAt,
          shareUrl,
          hubOptIn,
        ]
      properties:
        shareId:
          type: string
          pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
        status: { type: string, enum: [active, revoked, takedown] }
        snapshotVersion: { type: integer, minimum: 1 }
        title: { type: string }
        repositoryCount: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
        shareUrl:
          { type: string, description: relative SPA route for the share link }
        hubOptIn: { type: boolean }
...